Refactor the code in NameGenerationUtil
